Rename pkg/ -> internal/
